pyproject.toml
[tool.poetry]
name = "langchain-rag-tutorial-basic"
version = "0.1.0"
description = ""
authors = ["gmcosta <geu_costa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
langchain = "^0.3.10"
langchain-community = "^0.3.10"
langchain-openai = "^0.2.11"
unstructured = {extras = ["md"], version = "^0.16.9"}
chromadb = "^0.5.23"
openai = "^1.57.0"
tiktoken = "^0.8.0"
langchain-chroma = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
